rename custom conf folder
