+++ after/plugin/colors.lua
@@ -1,9 +1,9 @@
-require('rose-pine').setup({
+require('catppuccin').setup({
     disable_background = true
 })
 
-function ColorMyPencils(color) 
-	color = color or "rose-pine"
+function ColorMyPencils(color)
+	color = color or "catppuccin"
 	vim.cmd.colorscheme(color)
 
 	vim.api.nvim_set_hl(0, "Normal", { bg = "none" })
