+++ init.lua
@@ -1,2 +1,1 @@
-require("theprimeagen")
-
+require("kakashi")
